backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
pydantic
pydantic-settings
huggingface-hub
sentence-transformers
faiss-cpu
python-multipart
PyPDF2
python-docx
aiofiles
jinja2
python-jose
pymysql
cryptography
duckduckgo-search
httpx

